Moved VolumeControlButton to a new file and set default volume
