+++ frontend/js/tests/common/brainzplayer/SpotifyPlayer.test.tsx
@@ -28,7 +28,6 @@ const defaultContext = {
 const props = {
   refreshSpotifyToken: new APIService("base-uri").refreshSpotifyToken,
   show: true,
-  volume: 100,
   playerPaused: false,
   onPlayerPausedChange: (paused: boolean) => {},
   onProgressChange: (progressMs: number) => {},
